+++ torch_utils/torch_io.py
@@ -1,0 +1,20 @@
+import torch
+import os
+
+def save_model(epoch, model,
+               optimizer, path='./ckpt'):
+    torch.save({
+            'epoch': epoch,
+            'model_state_dict': model.state_dict(),
+            'optimizer_state_dict': optimizer.state_dict()
+            }, path)
+
+
+def load_model(model, epoch,
+               optimizer, path='./ckpt'):
+    if os.path.exists(path):
+        checkpoint = torch.load(path)
+        model.load_state_dict(checkpoint['model_state_dict'])
+        optimizer.load_state_dict(checkpoint['optimizer_state_dict'])
+        epoch[0] = checkpoint['epoch'] + 1
+    return
